Prefix template variable with @
Starting in Puppet 3.2. All variables in a compiled
erb template that access puppet variables from the scope
where the template was compiled should be prefixed with
@ to make them instance variables as opposed to local
variable/methods.

Failure to do so results in annoying and ugly deprecation
warnings.

The change was made b/c the ruby syntax for methods is the
same as local variables, and often lead to issues when a variable
name conflicted with built-in methods.

Referring to variables as instance variables in compatibile with
Puppet 2.7.x.