Don't restart carbon-cache to avoid data-loss
Daniel Werdermann informed me, "that a restart of the carbon cache means
losing data. And on bigger instances it can take minutes to restart a
carbon cache. The problem is that carbon-cache does not provide a proper
sighup command for logrotation."

Therefore, I removed the restart command. As the logfile is handled with
copytruncate, the daemon can keep the file and no restart is needed.