Handle permissions on /opt/graphite/storage more gently
Running `chown -R` on /opt/graphite/storage/whisper can take a very long
time on large installations with hundreds of thousands of metrics, and
it will also disrupt carbon-cache write operations if it is running with
a different user than root or web user. So I replaced it with an
explicit list of directories, which should be sufficient.

Also, most requirements on Exec['Chown graphite for web user'] were not
necessary, so I removed them, and replaced the others with requirements
on /opt/graphite/storage.