Small fixes and improvements (#281)
* Fixes that weren't properly merged

- The `wsgi.py` link is no longer needed since one was made in
  `config.pp`. So that code is no longer necessary. This block wasn't
  merged in.
- Removing subscription to `$local_settings_py_file`. It is not defined
  when `config_gunicorn.pp` is evaluated. As it turns out, it can be safely
  removed.

* New variable means WorkingDirectory doesn't need /graphite on the end

* Only run race condition fix after initial db creation

It used to `exec` on every run, now it won't.

* Only bump systemd on install

* Add service provider definition for Debian-based OSs

Debian 8 was not starting carbon-cache because it didn't know it had to
reload the systemd definitions. So I coded in a params for it, like was
done for redhat-based OSs. This allowed for simplified logic in
`config.pp`.

* Adding provider to gunicorn service

Ubuntu 15.10 has some weird race condition or something, where the
service would sometimes start and sometimes not. Explicitly specifying
the service provider seems to have no such error.

* Set RedHat service provider based on OS version

Rather than blanketing everything with RedHat and using OS version
checks, I have coded in what the `service_provider` should be.

* Adding rspec for config_gunicorn and updating the others

* Fix for float vs string problem