fix graphite.db permissions & location & remove default apache config
if installing from packages you get a default apache config -
graphite_web.conf. However, that has data in it which stops httpd from
starting therefore ensuring it's removed here as this module creates a
graphite.conf.

changing the parameter used for the graphite.db location to ensure that
the permissions are set on the correct.