Require pip before installing Django
The provider for the Django package is pip by default, but the Django package
resource currently isn't set to require `pip`, which can lead to an error when
pip is not already installed on the target system and Puppet decides to try to
install Django before installing pip. This commit fixes that issue.