Symlink correct target when installing carbon or gweb from a URL or file source
Previously, setting `gr_carbon_source` or `gr_graphite_source` to a URL or local
file representing the appropriate pip package, while also setting the
coresponding package version to `'present'` would result in Puppet attempting to
use names like `graphite_web-present-py2.7.egg-info` when creating symlinks to
the packages. This fixes that so that the correct file names are used.