Update carbon-logrotate - ignore unrotated files
By default carbon-logrotate will attempt to compress files that haven't been rotated or updated by carbon in over 3 days. This will eventually begin to fail as it cannot gzip over the same name (we see this happening most often on listener.log)

This seems to happen because carbon doesn't do the file rotation until the next line is added to the log. 

This change will only attempt to compress files that carbon has actually rotated

Solves #371